Setup for the first big run
ran it on AWS with first 1000, then 5000 then 2500 executor.queueSize limit. The
higher limits caused input/output errors on /s3mnt, with goofys, or the
container could not be started at all, with a werid error: error while
creating mount source path '/s3mnt': mkdir /s3mnt, file exists.

Possible workarounds/solutions:
- Limit job submission rate via executor.submitRateLimit, eg '100/2min',
100 jobs every 2 minutes.
- Increase temp space on VM, and copy the files for processing there.
- umount/mount bucket, make multiple tests if IO works, and sleep between
attempts.

Also, there were some errors for out of memory during the last step of
compression, thus zstd needs higher memory for the next run.

A snapshot of the status:
[52/c0d68f] process > CloudDiamond:download_sra (ERR3483308)           [100%] 43350 of 43350, cached: 43243, failed: 85 ✔
[52/9403b4] process > CloudDiamond:convert_sra_to_fastq (SRR14139965)  [ 99%] 43219 of 43265, cached: 1886, failed: 21625
[15/544df9] process > CloudDiamond:join_paired_end_reads (ERR9708723)  [ 85%] 18477 of 21594, cached: 245, failed: 10608
executor >  awsbatch (64563)
[52/c0d68f] process > CloudDiamond:download_sra (ERR3483308)           [100%] 43350 of 43350, cached: 43243, failed: 85 ✔
executor >  awsbatch (82701)
[52/c0d68f] process > CloudDiamond:download_sra (ERR3483308)          [100%] 43350 of 43350, cached: 43243, failed: 85 ✔
[87/66a287] process > CloudDiamond:convert_sra_to_fastq (SRR12712406) [100%] 43265 of 43265, cached: 1886, failed: 21628 ✔
[f0/2212f8] process > CloudDiamond:join_paired_end_reads (ERR9661576) [100%] 21637 of 21637, cached: 245, failed: 10752 ✔
[8e/b1c905] process > CloudDiamond:diamond_fastqs (SRR12229366)       [ 95%] 10375 of 10885, cached: 7, failed: 1430
[b3/11a339] process > CloudDiamond:compress_outputs (SRR12229366)     [ 99%] 8937 of 8945, failed: 3072

A problem was also that the download_sra step was the first to be
executed, thus flooding the bucket. Possible workarounds would be to
create multiple queues for each step, and set the later steps to
a higher priority. Alternatively, the workflow of GEMMaker could be
adopted.

+++ bin/process_accession.bash
@@ -1,5 +1,6 @@
 #!/bin/bash
-set -xv
+#set -xv
+set -o pipefail
 # This script gets the SRA accession and the aligner name as input parameters
 # It downloads the SRA file, converts it to fastq, aligns/diamondn's it to the reference genome
 # The output are stats files
@@ -17,9 +18,11 @@ echo "$(date) Processing accession $1"
 SRA_DIR=/s3mnt/scratch
 SRA_BUCKET=s3://diamond2-bucket1/scratch
 DIAMOND_DB=/data/ref/Col-CEN/ColCEN_ATHILA.dmnd
+PATH=${PATH}:/home/ec2-user/micromamba/envs/base/bin/:/home/ec2-user/sratoolkit.3.0.7-centos_linux64/bin
+DELETE_PREV=1 # Delete data from previous steps. 3rd step joined fastq is not deleted, compressed later!
 
 # Check if the number of arguments is correct
-if [ $# -lt 2 ] ; then
+if [ $# -lt 1 ] ; then
     echo "Usage: process_accession.bash <SRA accession> [step]"
     exit 1
 fi
@@ -28,7 +31,10 @@ fi
 SRA_ACCESSION=$1
 STEP=$2
 
+DOWNLOAD_DIR=$SRA_DIR/$SRA_ACCESSION/1_download
 FASTQ_DIR=$SRA_DIR/$SRA_ACCESSION/2_fastq
+JOIN_DIR=$SRA_DIR/$SRA_ACCESSION/3_join
+BLAST_DIR=$SRA_DIR/$SRA_ACCESSION/4_diamondn
 
 ##########################################################################################
 # 1. Download the SRA file
@@ -36,7 +42,6 @@ FASTQ_DIR=$SRA_DIR/$SRA_ACCESSION/2_fastq
 if [[ "${STEP}" == "" || "${STEP}" == "1_download" ]] ; then 
 echo "$(date) 1. Download the SRA file"
 # Create an accession directory
-DOWNLOAD_DIR=$SRA_DIR/$SRA_ACCESSION/1_download
 mkdir -p ${DOWNLOAD_DIR}
 
 # Download the SRA file
@@ -44,7 +49,7 @@ mkdir -p ${DOWNLOAD_DIR}
 # Check if the download was already done and successful (the flag file exists)
 if [ ! -f ${SRA_DIR}/${SRA_ACCESSION}/1_download.done ]; then
     # Download the SRA file via AWS CLI
-    aws s3 cp s3://sra-pub-run-odp/sra/${SRA_ACCESSION}/${SRA_ACCESSION} ${SRA_BUCKET}/${SRA_ACCESSION}/1_download/
+    /home/ec2-user/micromamba/envs/base/bin/aws s3 cp s3://sra-pub-run-odp/sra/${SRA_ACCESSION}/${SRA_ACCESSION} ${SRA_BUCKET}/${SRA_ACCESSION}/1_download/
 
     # Check if the download was successful
     if [ $? -ne 0 ]; then
@@ -66,7 +71,7 @@ echo "$(date) 2. Convert the SRA file to fastq(s)"
 mkdir -p ${FASTQ_DIR}
 
 # Check if the conversion was already done and successful (the flag file exists)
-if [ ! -f ${SRA_DIR}/${SRA_ACCESSION}/2_fastq.done ]; then
+if [ ! -f ${SRA_DIR}/${SRA_ACCESSION}/2_fastq.done ] && [ -f ${SRA_DIR}/${SRA_ACCESSION}/1_download.done ] ; then
     CURR_DIR=$(pwd)
     cd ${FASTQ_DIR}
     # Convert the SRA file to fastq(s)
@@ -82,8 +87,14 @@ if [ ! -f ${SRA_DIR}/${SRA_ACCESSION}/2_fastq.done ]; then
         exit 1
     fi
     cd ${CURR_DIR}
+
+    # Delete previous step data
+    if [[ "${DELETE_PREV}" == "1" ]] ; then
+        rm ${DOWNLOAD_DIR}/${SRA_ACCESSION}
+    fi
+
     # Create the flag file
-    touch -f ${SRA_DIR}/{$SRA_ACCESSION}/2_fastq.done
+    touch -f ${SRA_DIR}/${SRA_ACCESSION}/2_fastq.done
 fi
 fi # STEP
 
@@ -99,7 +110,7 @@ if [[ -f ${FASTQ_DIR}/${SRA_ACCESSION}_1.fastq ]] && [[ -f ${FASTQ_DIR}/${SRA_AC
     mkdir -p ${JOIN_DIR}
 
     # Check if the joining was already done and successful (the flag file exists)
-    if [ ! -f ${SRA_DIR}/${SRA_ACCESSION}/3_join.done ]; then
+    if [ ! -f ${SRA_DIR}/${SRA_ACCESSION}/3_join.done ] && [ -f ${SRA_DIR}/${SRA_ACCESSION}/2_fastq.done ] ; then
         # Join the paired-end reads
         fastq-join ${FASTQ_DIR}/${SRA_ACCESSION}_1.fastq ${FASTQ_DIR}/${SRA_ACCESSION}_2.fastq -o ${JOIN_DIR}/${SRA_ACCESSION}_%.fastq
         # Check if the joining was successful
@@ -121,7 +132,7 @@ else # Just copy the fastq file to the join directory
     mkdir -p ${JOIN_DIR}
 
     # Check if the joining was already done and successful (the flag file exists)
-    if [ ! -f ${SRA_DIR}/${SRA_ACCESSION}/3_join.done ]; then
+    if [ ! -f ${SRA_DIR}/${SRA_ACCESSION}/3_join.done ] && [ -f ${SRA_DIR}/${SRA_ACCESSION}/2_fastq.done ] ; then
         # Copy the fastq file to the join directory
         cp ${FASTQ_DIR}/${SRA_ACCESSION}.fastq ${JOIN_DIR}/${SRA_ACCESSION}.merged.fastq
 
@@ -134,6 +145,12 @@ else # Just copy the fastq file to the join directory
         # Create the flag file
         touch -f ${SRA_DIR}/${SRA_ACCESSION}/3_join.done
     fi
+
+    # Delete data from previous step
+    if [[ "${DELETE_PREV}" == "1" ]] ; then
+        rm ${FASTQ_DIR}/${SRA_ACCESSION}*.fastq
+    fi
+
 fi
 fi # STEP
 
@@ -144,32 +161,145 @@ fi # STEP
 if [[ "${STEP}" == "" || "${STEP}" == "4_diamondn" ]] ; then 
 echo "$(date) 4. Blast the fastq(s) against the reference genome"
 # Create a directory for the blast results
-BLAST_DIR=$SRA_DIR/$SRA_ACCESSION/4_diamondn
 mkdir -p ${BLAST_DIR}
 
 # Check if the blast was already done and successful (the flag file exists)
 CURR_DIR=$(pwd)
-if [ ! -f ${SRA_DIR}/${SRA_ACCESSION}/4_diamondn.done ]; then
+if [ ! -f ${SRA_DIR}/${SRA_ACCESSION}/4_diamondn.done ] && [ -f ${SRA_DIR}/${SRA_ACCESSION}/3_join.done ] ; then
     # Blast the fastq(s) against the reference genome
     cd ${BLAST_DIR}
 
-    diamond blastn --db ${DIAMOND_DB} -q ${SRA_DIR}/${SRA_ACCESSION}/3_join/${SRA_ACCESSION}.merged.fastq -o ${BLAST_DIR}/${SRA_ACCESSION}.diamondn \
-            -f 6 qseqid sseqid pident qstart qend sstart send evalue bitscore qlen slen \
-            --shape-mask 11111111111 \
-            --minimizer-window 1 \
-            --block-size 1.0 \
-            -k0 \
-            --evalue 0.0001 \
-            
-        
-    # Check if the blast was successful
-    if [ $? -ne 0 ]; then
-        echo "Error: diamondn failed for ${SRA_ACCESSION}"
-        exit 1
+    # 1. ColCEN
+    OUT_DIR=${BLAST_DIR}/ColCEN_ATHILA
+    mkdir -p ${OUT_DIR}
+    if [[ ! -f ${OUT_DIR}/4a_ColCEN_ATHILA.DONE ]] ; then
+        diamond blastn --db ${DIAMOND_DB} -q ${SRA_DIR}/${SRA_ACCESSION}/3_join/${SRA_ACCESSION}.merged.fastq -o ${OUT_DIR}/${SRA_ACCESSION}.diamondn \
+                -f 6 qseqid sseqid pident qstart qend sstart send evalue bitscore qlen slen cigar \
+                --shape-mask 11111111111 \
+                --minimizer-window 1 \
+                --block-size 1.0 \
+                --masking 0 \
+                -k0 \
+                --evalue 0.0001 
+        # Check if the blast was successful
+        if [ $? -ne 0 ]; then
+            echo "Error: diamondn failed for ${SRA_ACCESSION}"
+            exit 1
+        fi
+        touch -f ${OUT_DIR}/4a_ColCEN_ATHILA.DONE
     fi
+
+    # 2. Ath-Rep
+    OUT_DIR=${BLAST_DIR}/Ath-Rep
+    mkdir -p ${OUT_DIR}
+    if [[ ! -f ${OUT_DIR}/4b_Ath-Rep.DONE ]] ; then
+        diamond blastn --db /data/ref/Ath-Rep/Ath-Rep.dmnd -q ${SRA_DIR}/${SRA_ACCESSION}/3_join/${SRA_ACCESSION}.merged.fastq -o ${OUT_DIR}/${SRA_ACCESSION}.diamondn \
+                -f 6 qseqid sseqid pident qstart qend sstart send evalue bitscore qlen slen cigar \
+                --shape-mask 11111111111 \
+                --minimizer-window 1 \
+                --block-size 1.0 \
+                --masking 0 \
+                -k0 \
+                --evalue 0.001 
+        # Check if the blast was successful
+        if [ $? -ne 0 ]; then
+            echo "Error: diamondn failed for ${SRA_ACCESSION}"
+            exit 1
+        fi
+        touch -f ${OUT_DIR}/4b_Ath-Rep.DONE
+    fi
+    # 3. Col0-Telib
+    OUT_DIR=${BLAST_DIR}/Col0-Telib
+    mkdir -p ${OUT_DIR}
+    if [[ ! -f ${OUT_DIR}/4c_Col0-Telib.DONE ]] ; then
+        diamond blastn --db /data/ref/Col0-Telib/Col0-Telib.dmnd -q ${SRA_DIR}/${SRA_ACCESSION}/3_join/${SRA_ACCESSION}.merged.fastq -o ${OUT_DIR}/${SRA_ACCESSION}.diamondn \
+                -f 6 qseqid sseqid pident qstart qend sstart send evalue bitscore qlen slen cigar \
+                --shape-mask 11111111111 \
+                --minimizer-window 1 \
+                --block-size 1.0 \
+                -k0 \
+                --evalue 0.001 \
+                --masking 0 
+        # Check if the blast was successful
+        if [ $? -ne 0 ]; then
+            echo "Error: diamondn failed for ${SRA_ACCESSION}"
+            exit 1
+        fi
+        touch -f ${OUT_DIR}/4c_Col0-Telib.DONE
+    fi
+
     cd ${CURR_DIR}
+
+    # Delete data from previous step
+    #if [[ "${DELETE_PREV}" == "1" ]] ; then
+    #    rm ${JOIN_DIR}/${SRA_ACCESSION}*.fastq
+    #fi
+
     # Create the flag file
     touch -f ${SRA_DIR}/${SRA_ACCESSION}/4_diamondn.done
 fi
 echo "$(date) Done processing accession $1"
 fi # 4 STEP
+
+
+##########################################################################################
+# 5.  Compress outputs
+##########################################################################################
+if [[ "${STEP}" == "" || "${STEP}" == "5_compress" ]] ; then 
+echo "$(date) 5. Compress output and fastq merged input"
+
+# Check if the compression was already done and successful
+CURR_DIR=$(pwd)
+if [ ! -f ${SRA_DIR}/${SRA_ACCESSION}/5_compress.done ] && [ -f ${SRA_DIR}/${SRA_ACCESSION}/4_diamondn.done ] ; then
+ 
+    # Compress merged fastq
+    if [[ ! -f ${SRA_DIR}/${SRA_ACCESSION}/3_join/5a_compress.DONE ]] ; then
+        cat ${SRA_DIR}/${SRA_ACCESSION}/3_join/${SRA_ACCESSION}.merged.fastq | zstd -T0 --adapt -c - > ${SRA_DIR}/${SRA_ACCESSION}/3_join/${SRA_ACCESSION}.merged.fastq.zst
+        if [ $? -ne 0 ]; then
+            echo "Error: compression failed for ${SRA_ACCESSION} fastq"
+            exit 1
+        fi
+        touch -f ${SRA_DIR}/${SRA_ACCESSION}/3_join/5a_compress.DONE
+        rm ${SRA_DIR}/${SRA_ACCESSION}/3_join/${SRA_ACCESSION}.merged.fastq
+    fi
+
+    # Compress diamond  fastq
+    if [[ ! -f ${BLAST_DIR}/ColCEN_ATHILA/5b_compress.DONE ]] ; then
+        cat ${BLAST_DIR}/ColCEN_ATHILA/${SRA_ACCESSION}.diamondn | zstd -T0 --adapt -c - > ${BLAST_DIR}/ColCEN_ATHILA/${SRA_ACCESSION}.diamondn.zst
+        if [ $? -ne 0 ]; then
+            echo "Error: compression failed for ${SRA_ACCESSION} ColCEN_ATHILA"
+            exit 1
+        fi
+        touch -f ${BLAST_DIR}/ColCEN_ATHILA/5b_compress.DONE
+        rm ${BLAST_DIR}/ColCEN_ATHILA/${SRA_ACCESSION}.diamondn
+    fi
+
+    # Compress diamond  fastq
+    if [[ ! -f ${BLAST_DIR}/Ath-Rep/5c_compress.DONE ]] ; then
+        cat ${BLAST_DIR}/Ath-Rep/${SRA_ACCESSION}.diamondn | zstd -T0 --adapt -c - > ${BLAST_DIR}/Ath-Rep/${SRA_ACCESSION}.diamondn.zst
+        if [ $? -ne 0 ]; then
+            echo "Error: compression failed for ${SRA_ACCESSION} Ath-Rep"
+            exit 1
+        fi
+        touch -f ${BLAST_DIR}/Ath-Rep/5c_compress.DONE
+        rm ${BLAST_DIR}/Ath-Rep/${SRA_ACCESSION}.diamondn
+    fi
+
+    # Compress diamond  fastq
+    if [[ ! -f ${BLAST_DIR}/Col0-Telib/5d_compress.DONE ]] ; then
+        cat ${BLAST_DIR}/Col0-Telib/${SRA_ACCESSION}.diamondn | zstd -T0 --adapt -c - > ${BLAST_DIR}/Col0-Telib/${SRA_ACCESSION}.diamondn.zst
+        if [ $? -ne 0 ]; then
+            echo "Error: compression failed for ${SRA_ACCESSION} Col0-Telib"
+            exit 1
+        fi
+        touch -f ${BLAST_DIR}/Col0-Telib/5d_compress.DONE
+        rm ${BLAST_DIR}/Col0-Telib/${SRA_ACCESSION}.diamondn
+    fi
+    
+    
+    # Create the flag file
+    touch -f ${SRA_DIR}/${SRA_ACCESSION}/5_compress.done
+
+fi # .done
+
+fi # Step
